requirements.txt
pillow>=10.0.0
pillow-heif>=0.15.0
tqdm>=4.66.0
tkinterdnd2>=0.3.0
